컴포넌트 단순화 및 이름 규칙 통일
