+++ src/components/EditorModal/ModalStyle.js
