+++ src/components/EditorModal/PutModal.jsx
@@ -8,15 +8,26 @@ import { POST_REQUEST } from 'reducers/post';
 import authHeader from 'sagas/auth-header';
 import axios from 'axios';
 
-import * as Styled from './PostModalStyle';
+import * as Styled from './ModalStyle';
 
-const PostModal = ({ onClose, maskClosable, closable, visible, titleText, postContent }) => {
+const PutModal = ({
+  onClose,
+  maskClosable,
+  closable,
+  visible,
+  titleText,
+  postContent,
+  tags,
+  onChangeTags,
+  thumbnailUrl,
+  onChangeThumbnail,
+}) => {
   const navigate = useNavigate();
   const dispatch = useDispatch();
   const { postError, postDone } = useSelector((state) => state.post);
 
-  const [tags, setTags] = useState([]);
-  const [thumbnailUrl, setThumbnailUrl] = useState('');
+  //   const [tags, setTags] = useState([]);
+  //   const [thumbnailUrl, setThumbnailUrl] = useState('');
   const [imageError, setimageError] = useState('');
   const [errorAnimation, setErrorAnimation] = useState(false);
 
@@ -40,16 +51,19 @@ const PostModal = ({ onClose, maskClosable, closable, visible, titleText, postCo
   const onImageChange = (e) => {
     (async () => {
       const img = e.target.files[0];
+      console.log(img);
       const formData = new FormData();
       formData.append('image', img);
       try {
         const { data: filename } = await axios.post('/articles/upload_image/', formData, {
           headers: authHeader(),
         });
+        onChangeThumbnail(`${filename.url}`);
+        // setThumbnailUrl(`${filename.url}`);
         setimageError('');
-        setThumbnailUrl(`${filename.url}`);
       } catch (err) {
-        setThumbnailUrl('');
+        onChangeThumbnail('');
+        // setThumbnailUrl('');
         setimageError('올바르지 않은 파일 형식입니다.');
         setErrorAnimation(false);
       }
@@ -57,8 +71,15 @@ const PostModal = ({ onClose, maskClosable, closable, visible, titleText, postCo
     return false;
   };
 
+  useEffect(() => {
+    console.log(thumbnailUrl);
+    // inputRef.current.files = thumbnailUrl;
+  }, []);
+
   const onTagsChange = (e) => {
-    setTags(e.target.value.split(','));
+    console.log(e);
+    onChangeTags(e.target.value.split(','));
+    // setTags(e.target.value.split(','));
   };
 
   const onSubmitPost = useCallback(() => {
@@ -145,4 +166,4 @@ const PostModal = ({ onClose, maskClosable, closable, visible, titleText, postCo
   );
 };
 
-export default PostModal;
+export default PutModal;
